Moved node startup to main_gr()

+++ node.sh
@@ -12,33 +12,42 @@ fi
 echo "==> Ensuring .bashrc exists and is writable"
 touch ~/.bashrc
 
-echo "==> Installing node version manager (NVM). Version $INSTALL_NVM_VER"
-# Removed if already installed
-rm -rf ~/.nvm
-# Unset exported variable
-export NVM_DIR=
+# Check if node is already installed
+if command -v node &> /dev/null; then
+	echo "==> Node.js is already installed. Skipping installation."
+	node --version
+	npm --version
+else
+	echo "==> Installing node version manager (NVM). Version $INSTALL_NVM_VER"
+	# Removed if already installed
+	rm -rf ~/.nvm
+	# Unset exported variable
+	export NVM_DIR=
 
-# Install nvm 
-curl -o- https://raw.githubusercontent.com/creationix/nvm/v$INSTALL_NVM_VER/install.sh | bash
-# Make nvm command available to terminal
-source ~/.nvm/nvm.sh
+	# Install nvm 
+	curl -o- https://raw.githubusercontent.com/creationix/nvm/v$INSTALL_NVM_VER/install.sh | bash
+	# Make nvm command available to terminal
+	source ~/.nvm/nvm.sh
 
-echo "==> Installing node js version $INSTALL_NODE_VER"
-nvm install $INSTALL_NODE_VER
+	echo "==> Installing node js version $INSTALL_NODE_VER"
+	nvm install $INSTALL_NODE_VER
 
-echo "==> Make this version system default"
-nvm alias default $INSTALL_NODE_VER
-nvm use default
+	echo "==> Make this version system default"
+	nvm alias default $INSTALL_NODE_VER
+	nvm use default
 
+	echo "==> Checking for versions"
+	nvm --version
+	node --version
+	npm --version
 
-echo "==> Checking for versions"
-nvm --version
-node --version
-npm --version
+	echo "==> Print binary paths"
+	which npm
+	which node
 
-echo "==> Print binary paths"
-which npm
-which node
+	nvm cache clear
+fi
 
-nvm cache clear
 npm install
+
+node index.js
